create method for generating form groups of type NutrientGroupType
