+++ libs/apps/ingredient-manager/src/lib/ingredient-manager/views/ingredient-editor/form-generator/primatives/multi-value/element-weight-group/element-weight-group.ts
@@ -1,17 +1,7 @@
 import { FormControl, FormGroup } from "@angular/forms";
-import { ElementWeightType, WeightType } from "@bake-a-weigh/site-types";
-
-
-
-
-
-export interface ElementWeightGroupType{
-
-    amount      : FormControl< number | null >,
-    weightType  : FormControl< WeightType | null >,
-    percentage? : FormControl< number | null >
-
-}
+import { ElementWeightType } from "@bake-a-weigh/site-types";
+import { ElementWeightGroupType } from "../../../form-types";
+import { createWeightControl } from "../../single-value/measurements/weight-control/weight-control";
 
 
 
@@ -22,7 +12,7 @@ export function createElementWeightGroup( defaultValues? : ElementWeightType ) :
     const group : FormGroup< ElementWeightGroupType > = new FormGroup< ElementWeightGroupType >({
 
         amount     : new FormControl< number | null >( defaultValues ? defaultValues.amount : null ),
-        weightType : new FormControl< WeightType | null >( defaultValues ? defaultValues.weightType : null ),
+        weightType : createWeightControl( defaultValues ? defaultValues.weightType : undefined ),
         percentage : ( defaultValues?.percentage ? new FormControl< number | null >( defaultValues.percentage ) : undefined )
 
     });
