+++ libs/apps/ingredient-manager/src/lib/ingredient-manager/views/ingredient-editor/form-generator/primatives/multi-value/ingredient-note-group/ingredient-note-group.ts
@@ -1,11 +1,6 @@
 import { FormControl, FormGroup } from "@angular/forms";
 import { IngredientNote }         from "@bake-a-weigh/site-types";
-
-
-export interface IngredientNoteFormGroup{
-    title       : FormControl< string >,
-    description : FormControl< string >
-}
+import { IngredientNoteFormGroup } from "../../../form-types";
